Fix token parser structure to be more robust.
Demand that the response is in the expected form: root object is an
array, containing one wrapper object, which has a "trends" key.
Previously, the parser would just look for a "trends" string followed by
an array, which was not correct.

Also add a skip state, which is used to ignore unwanted object values
with children (i.e. arrays and objects). This ensures that the token
counter approach doesn't fall over if the "trend" objects have
additional key-value pairs added which have more children.