Fix faulty loop initialisation.
j should be set to 1, meaning "traverse the next 1 token". In the loop,
the number of children of the current token are added to j, and at the
end of each loop j is decremented.

Therefore for a JSON string with an empty array or object, the loop will
run once. For other results, it will cover all tokens.